allow floats for cliff time

+++ src/app/genesis_ledger_from_tsv/genesis_ledger_from_tsv.ml
@@ -128,8 +128,8 @@ let validate_fields ~wallet_pk ~amount ~initial_min_balance ~cliff_time_months
   in
   let valid_cliff_time_months =
     try
-      let n = Int.of_string cliff_time_months in
-      n >= 0
+      let n = Float.of_string cliff_time_months in
+      n >= 0.0
     with _ -> false
   in
   let valid_cliff_amount = valid_mina_amount cliff_amount in
